Move notification helper to its own feature package.
